% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{final_period_date}
\alias{final_period_date}
\title{final_period_date
This function produces the second thursday of the next quarter}
\usage{
final_period_date(date_val, month_str, year_val)
}
\arguments{
\item{date_val}{A number}

\item{month_str}{A word in full word fomat}

\item{year_val}{A number}
}
\description{
final_period_date
This function produces the second thursday of the next quarter
}
\examples{

}
