% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{incrs_or_decrs}
\alias{incrs_or_decrs}
\title{Outputs a decriptive word or phrase that decribes the CI that is inputted}
\usage{
incrs_or_decrs(no1, no2, type)
}
\arguments{
\item{no1}{A number}

\item{type}{Shows what type of stat is being inputted}
}
\description{
Outputs a decriptive word or phrase that decribes the CI that is inputted
}
\examples{
format_num(5,10,"days") # returns "increases the time to first reoffence"
format_num(-5,10,"days") # returns "increases/decreases/has no effect on the time to first reoffence"
}
