% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralising_functions.R
\name{people_func}
\alias{people_func}
\title{This function outputs the correct singular or plural version of person, man or woman}
\usage{
people_func(value, gender)
}
\arguments{
\item{value}{A number}

\item{gender}{A word}
}
\description{
This function outputs the correct singular or plural version of person, man or woman
}
\examples{
people_func(1,'people') #returns "person"
people_func(3,'female') #returns "women"
}
