% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_graph.R
\name{time_graph}
\alias{time_graph}
\title{This function outputs the 'time before first proven reoffence' graph}
\usage{
time_graph(data, text_size, ci_text_size = 5, y_title_size = 11,
  legend_text_size = 12, comp_CI_position = 0, treat_CI_position = 0)
}
\arguments{
\item{data}{The standard result dataset}

\item{text_size}{Generic text size in graph}

\item{ci_text_size}{Text size of confidence intervals}

\item{y_title_size}{Text size of y-axis label}

\item{legend_text_size}{Text size of legend text}

\item{comp_CI_position}{Horizontal position of the comparison group CI label with respect to the x-axis}

\item{reoff_CI_position}{Horizontal position of the treatment group CI label with respect to the x-axis}
}
\description{
This function outputs the 'time before first proven reoffence' graph
}
